tests/wpt/web-platform-tests/XMLHttpRequest/event-timeout.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: timeout event</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#handler-xhr-ontimeout" data-tested-assertations="../.." />
    <link rel="help" href="https://xhr.spec.whatwg.org/#event-xhr-timeout" data-tested-assertations="../.." />
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-timeout-attribute" data-tested-assertations="following-sibling::ol/li[2]" />
    <link rel="help" href="https://xhr.spec.whatwg.org/#timeout-error" data-tested-assertations=".." />
    <link rel="help" href="https://xhr.spec.whatwg.org/#infrastructure-for-the-send()-method" data-tested-assertations="following-sibling::dl//code[contains(@title,'dom-XMLHttpRequest-timeout')]/.. following-sibling::dl//code[contains(@title,'dom-XMLHttpRequest-timeout')]/../following-sibling::dd" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test();
      test.step(function() {
        var client = new XMLHttpRequest();
        client.ontimeout = function() {
          test.step(function() {
            assert_equals(client.readyState, 4);
            test.done();
          });
        };
        client.timeout = 5;
        client.open("GET", "resources/delay.py?ms=20000");
        client.send(null);
        test.step_timeout(() => {
          assert_unreached("ontimeout not called.");
        }, 10);
      });
    </script>
  </body>
</html>
